Moves all travel sound logic to its own component because it is a self contained system
