+++ Assets/Prefabs/Actors/Player.prefab
@@ -331,6 +331,7 @@ GameObject:
   m_Component:
   - 4: {fileID: 404084}
   - 95: {fileID: 9511250}
+  - 114: {fileID: 11449098}
   m_Layer: 0
   m_Name: Final_Idle
   m_TagString: Untagged
@@ -1552,6 +1553,7 @@ GameObject:
   - 114: {fileID: 11495812}
   - 114: {fileID: 11476420}
   - 114: {fileID: 11488222}
+  - 114: {fileID: 11455752}
   m_Layer: 10
   m_Name: Player
   m_TagString: Untagged
@@ -3111,8 +3113,8 @@ Transform:
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
   m_GameObject: {fileID: 180122}
-  m_LocalRotation: {x: 0, y: .800290287, z: 0, w: -.599612832}
-  m_LocalPosition: {x: 93.1800003, y: 50.8899994, z: 41.2000008}
+  m_LocalRotation: {x: 0, y: .440437436, z: 0, w: -.897783399}
+  m_LocalPosition: {x: 8.69999981, y: 52.2999992, z: -15.5}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
   - {fileID: 404084}
@@ -3850,11 +3852,32 @@ MonoBehaviour:
   _interactableLayer:
     serializedVersion: 2
     m_Bits: 131072
+--- !u!114 &11449098
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 119804}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 33c7efc44a47ee445ae908d8282d51d0, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &11455752
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 180122}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 99670a7f24d87ed4096195a9c82d8014, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
   _stepSoundSource: {fileID: 8247732, guid: d4a7e1fba3d1c5d459b36e08e4831bd1, type: 2}
   _groundStepSounds: {fileID: 11400000, guid: 7f2dac2b5f313bd4a8d495737112d48d, type: 2}
   _waterStepSounds: {fileID: 11400000, guid: a6fdfad4585b37d45b8b1ea929eb8968, type: 2}
   _brushStepSounds: {fileID: 11400000, guid: d5071f8fc891d1549a8820157f9283b9, type: 2}
-  stepTime: 1
 --- !u!114 &11476420
 MonoBehaviour:
   m_ObjectHideFlags: 1
