+++ Assets/Prefabs/Audio/SFX/Step Sound.prefab
@@ -37,7 +37,7 @@ AudioSource:
   m_Enabled: 1
   serializedVersion: 4
   OutputAudioMixerGroup: {fileID: 0}
-  m_audioClip: {fileID: 8300000, guid: 11dc13a6d48bd85459c3be43c39e1498, type: 3}
+  m_audioClip: {fileID: 8300000, guid: 0d32e49edebd5a64aa5ccaea4c12a94e, type: 3}
   m_PlayOnAwake: 1
   m_Volume: 1
   m_Pitch: 1
