+++ Assets/Scripts/Player/TravelSounds.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 99670a7f24d87ed4096195a9c82d8014
+timeCreated: 1430536472
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
